kernel lib: add implemention for num_to_string and num_counsts functions
The functions of converting numbers to strings take a pointer to array
of characters (allocated memory), number and the base of number and
write the result to arrays of character.

The functions of getting count of number take a number and the base
of number and return the count of number.

Signed-off-by: Maksym Prymierov <primerovmax@gmail.com>